+++ .vscode/settings.json
@@ -62,7 +62,27 @@
     "editor.formatOnSave": true,
     "terminal.integrated.persistentSessionReviveProcess": "never",
     "workbench.colorCustomizations": {
-        "activityBar.activeBorder": "#0bb45d"
+        "activityBar.activeBorder": "#0bb45d",
+        "activityBar.activeBackground": "#914b09",
+        "activityBar.background": "#914b09",
+        "activityBar.foreground": "#e7e7e7",
+        "activityBar.inactiveForeground": "#e7e7e799",
+        "activityBarBadge.background": "#0bb45d",
+        "activityBarBadge.foreground": "#e7e7e7",
+        "commandCenter.border": "#e7e7e799",
+        "editorGroup.border": "#914b09",
+        "panel.border": "#914b09",
+        "sash.hoverBorder": "#914b09",
+        "statusBar.background": "#613206",
+        "statusBar.foreground": "#e7e7e7",
+        "statusBarItem.hoverBackground": "#914b09",
+        "statusBarItem.remoteBackground": "#613206",
+        "statusBarItem.remoteForeground": "#e7e7e7",
+        "tab.activeBorder": "#914b09",
+        "titleBar.activeBackground": "#613206",
+        "titleBar.activeForeground": "#e7e7e7",
+        "titleBar.inactiveBackground": "#61320699",
+        "titleBar.inactiveForeground": "#e7e7e799"
     },
     "python.terminal.activateEnvInCurrentTerminal": true,
     "todo-tree.filtering.ignoreGitSubmodules": true,
